Reorganised second virial code for future optimisations.
Code was reorganised but not really changed.

Part of #108
